new dialog folder structure refactor
