+++ packages/teams-js/src/internal/appHelpers.ts
@@ -13,7 +13,7 @@ import { compareSDKVersions, inServerSideRenderingEnvironment, runWithTimeout } 
 import * as app from '../public/app/app';
 import * as authentication from '../public/authentication';
 import { FrameContexts } from '../public/constants';
-import { dialog } from '../public/dialog';
+import { dialog } from '../public/dialog/dialog';
 import { menus } from '../public/menus';
 import { pages } from '../public/pages';
 import {
